Occlusion + initial dialogue system commit
A commit of the initial work done for my dissertation. Contains numerous scenes that act as testing playgrounds as i develop the audio systems utilizing the fmod api.

Initial implementations involve 3D script with audio occlusion that utilizes line renders to emulate sound waves, with parameter control for the sound size/width so diffraction can be emulated and designer controlled through the inspector variable on the script.

Also the base implementation of the dialogue system has been implemented. Utilizing xml to allow users to design files that contain all lines for an NPC, along with keys to trigger the line. This is deseriazlized with a static utility script that the Entity script (attached to npc for respective dialogue triggers) will provide the xml file name for, located in streaming assets.
There are also multiple ways to trigger dialogue, such as collision or trigger enter / exit. UI inspector vars are overridden depending on which enum for trigger type the user selects, providing easy use of the system.