Added hashtables
Created the new data structure for handling dialogues, as well as linking it to game data. This allows game datas to be saved as an object type in a hashtable for game data. This is also similar for dialogue as it is deserialized from xml and then that populates an entitiy's different data objects into a entity object, comprised of: lines -> lineconditions. This allows an entity to contain all information about a line through various collections, which the entity is then stored in the NPChashtable