Added Dialogue Info callback + sequencing
Added a custom callback, similar to the one for triggering dialogue. However this one just uses a openonly sound mode, so its only for reading sound info, not triggering it. This allows us to get the length of a dialogue clip before it is triggered and then delay the trigger of the next line until it has concluded, aka sequencing.

The dialogue system is now fully capable of iterating through all lines and checking all conditions against the respective game data hash table values, checking condition and then either triggering or not.