Added sequence mode functionality
Added ability to be able to select a enum of the sequence type, currently random one shot and sequence.

Random one shot checks which lines of that node have true conditions and then plays a random line from that list.

Sequence mode sequentially plays the dialogue one after another. 

The user can select the type on the drop down on in the inspector for the script currently called testxml