Refactor + implemented full quest for dialogue scene
Refactored dialogue system: implemented the broadcast pattern where the dialogue manager is the subject and the different components of the dialogue system listen for invokes which are passed a dialogue state enum to determine behavior's for each dialogue component. This also provides the architecture for providing game designers with modular ability to trigger different things. On the dialogue entity script there are lists for different events that are part of the dialogue system: e.g. convo end or single line start.



Implemented a quest to highlight all different parts of dialogue system with conditions and branching